service-user/requirements.txt
flask
pyjwt
python-dotenv
flask-sqlalchemy
flask-migrate
werkzeug